requirements.txt
Django
django-cryptography-django5==2.2
python-decouple
psycopg2-binary
dj-database-url
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-allauth
dj-rest-auth
social-auth-app-django
django-modeltranslation
django-cloudinary-storage
cloudinary
stripe
django-redis
whitenoise
black
flake8
gunicorn
django-phonenumber-field[phonenumbers]


# testing! - remove in production
pytest 
pytest-django 
pytest-cov 
pytest-factoryboy
Faker
python-dateutil

